moved disk functions requiring libc to stdlib.c
